refactor: merge cloudflare deploy setup script
